feat(generateComponents): add util function
Add a function that takes an array of item creator functions as an argument and returns a function that takes an object as an argument that contains the data and the item number. The item creator is a function that returns an array of objects. The objects types can be based on ListItem, TableContent or ConfigItem. More types may be added in the future.

The array of objects returned by the item creator function are used as props for components that generate a list/group of components.