feat: Add functions to create card list items and dialog configurations
Add two functions to create card component list items and dialog configurations. The createCardListItems function takes an object with data and a component as arguments. If the data is an array, it maps over the array and returns an array of list items. If the data is not an array, it returns a single list item. The createDialogConfig function takes an object with data, content, and trigger as arguments. If the data is an array, it maps over the array and returns an array of dialog configurations. If the data is not an array, it returns a single dialog configuration. The implementation uses the cardListItem and dialogConfigItem functions to create the list items and dialog configurations.