feat(components): Add component for displaying featured products
Create FeaturedProducts component to be used for displaying a slider of featured products. The component takes a data prop, which receives An array of Products or Partial<Product>[] representing the featured products to be displayed and a hasReversedSlider prop which is an Optional boolean indicating whether the slider should be reversed.