feat(components): add a new component to render a gallery with pagination
Create Gallery component which can be used to render a gallery of products with pagination. The component takes an array of products as an input and renders a list of cards, each of which contains a product. The component also supports buttons to navigate to the previous and next page.

The component is implemented using the ReactPaginate library.