feat(components): Add ProductGallery component
Create Product Gallery component which contains the Gallery component, a Searchbar component and a category filter. The search bar allows users to search for products by name or description. The category filter allows users to filter products by category. The search functionality is being implemented by using the lunr library, which adds a search engine to the client side. This component also has a Loading component that is displayed while the data is being fetched.