feat(Navbar/helper): add useDataStateObserver hook
Create useDataStateObserver hook, to be used to listen for changes to the data-state attribute of an HTML element. The Radix UI NavigationMenu primitive and dropdown (or Sub menus) contain a data-state attribute. This is useful for components that need to be aware of changes to the data-state attribute. This is used to make the Navbar and containing dropdowns share the current data-sate with one another. This helps each component in the Navbar react to any changes made to another