feat(hoc): Add withScaleEffec HOC
This commit introduces a Higher Order Component (HOC) named 'withScaleEffect'. This HOC takes a component (specifically a button component) and returns a new component that applies a scale effect when the component is clicked or touched.

This HOC can  provide a visual feedback to the user that the component has been interacted with.